+++ packages/codemirror/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./createLineNumbers";
 export * from "./createReadOnly";
 export * from "./createWrapLine";
+export * from "./createExtensions";